refactor(#2605): split df_spike/tests.rs by responsibility (campsite rule)
The `--lite` file-size ratchet FAILed: `tests.rs` reached 836 lines against the
800-line source limit. It is a test file, but `df_spike/tests.rs` matches none of
the ratchet's test patterns (`*_tests.rs`, `*/tests/*`), so it is measured as
source — and at that size it was overdue for splitting anyway.

Split into `df_spike/tests/` by the property under test (#1116/#1135):

* `support.rs` — the two-generation fixture, the pinned read clock, and the
  row-rendering/arm-driving helpers, plus the shared imports re-exported so each
  file opens with one `use super::support::*;`
* `pushdown.rs` — `Expr` classification (`Exact` only for what the scan applies)
* `provider.rs` — the `TableProvider` surface: schema, projection, fail-closed open
* `equivalence.rs` — the arm-equivalence oracle and the `count(*)`-under-pushdown
  regression tests
* `harness.rs` — the harness's own contract: operand parsing, scenario/arm ids,
  the SQL each scenario runs, what a result record reports

No test was changed, added or removed: 24 before, 24 after, all passing. Largest
file is now 258 lines. The report's file table and test count are updated.

Refs #2605
